test: replace templates pages
